Hibernate create DB.  Add function to insert data.
