+++ AdoptionApp/src/main/java/ttl/project/App/AdoptionsController.java
@@ -1,12 +1,12 @@
-package ttl.Project.App;
+package ttl.project.App;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import ttl.Project.Domain.AdoptionProcess;
-import ttl.Project.Service.AdoptionService;
+import ttl.project.domain.AdoptionProcess;
+import ttl.project.Service.AdoptionService;
 
 import java.net.URI;
 import java.util.List;
