+++ AdoptionApp/src/main/java/ttl/project/domain/JDBCTemplate.java
@@ -1,14 +1,11 @@
-package ttl.Project.Domain;
+package ttl.project.domain;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class JDBCTemplate {
     public static void main(String[] args) {
