+++ AdoptionApp/src/main/resources/application.properties
@@ -1,2 +1,12 @@
 spring.application.name=AdoptionApp
 spring.profiles.active=dev
+
+################################Hibernate DDL ###############################
+#Does Hibernate generate the DDL (false for no)
+spring.jpa.generate-ddl=true
+#spring.jpa.generate-ddl=false
+#If Hibernate, then a further switch is
+# available to customize how Hibernate works
+#spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=create
+################################End Hibernate DDL ###############################
