+++ AdoptionApp/src/test/java/ttl/project/controller/AdoptionsControllerTest.java
@@ -1,4 +1,4 @@
-package ttl.Project.controller;
+package ttl.project.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
